feat(routes): set routes private

+++ routes/api.php
@@ -52,18 +52,17 @@ Route::group(['prefix' => 'v1/auth', 'middleware' => 'api'], function () {
     Route::post('password/reset', 'Auth\ResetPasswordController@reset');
 });
 
-Route::group(['prefix' => 'v1', 'middleware' => 'api'], function () {
+Route::group(['prefix' => 'v1', 'middleware' => 'jwt.auth'], function () {
     // ACL
     Route::group(['prefix' => 'acl'], function () {
         Route::apiResource('roles', 'RoleController');
         Route::apiResource('permissions', 'PermissionController');
         Route::apiResource('roles/{role}/permissions', 'RolePermissionController');
-        Route::delete('roles/{role}/permissions/{permission}', 'RolePermissionController@destroy');
+        Route::delete('roles/{role}/permissions/{permission}', 'RolePermissionController@destroy')->name('role_permissions.destroy');
     });
 
     // User
     Route::apiResource('users', 'UserController');
-    Route::put('users/{user}/approve', 'UserController@approve');
-    Route::put('users/{user}/block', 'UserController@block');
-    Route::put('users/{user}/role', 'UserController@role');
+    Route::put('users/{user}/approve', 'UserController@approve')->name('users.approve');
+    Route::put('users/{user}/block', 'UserController@block')->name('users.block');
 });
